Fix assembly errors that were introduced with the asm / orient bump (#1955)
63b3ba2b fixes the errors produced when parsing java9 byte
code. sbt-assembly began to fail, for the following reasons:

Orient brought in two different artifacts that provided the same library in
the same package paths:

- javax.xml.stream:stax-api:1.0-2 (published 2008)
- stax:stax-api:1.0.1 (published 2006)

Since javax.xml.stream:stax-api:1.0-2 looked more modern, we ignore the
stax:stax-api:1.0.1 dependency. Things seem to work fine with this and this
resolves a good chunk of the merge conflict errors produced by sbt-assembly

Further, asm 6 dependencies provide various `module-info.class` in the root
of the JAR. It is not possible to merge these so we just discard them.